Add the ability for the prCheck to comment back on github.
In order to use this, you'll need to generate an oauth token and pass it
in.  This curl call works:

curl -i -u YOUR_USERNAME -d '{"scopes": ["repo"]}' https://api.github.com/authorizations

It should hopefully be easy enough to build on this and get it automated
and even more useful.